fix(pesit): accept empty password in TLS mutual auth (cert replaces password)
In PeSIT-TLS with mutual authentication, some products (e.g. Axway CFT)
consider the validated client certificate as sufficient authentication
and send an empty password in PI 5. The Gateway was rejecting these
connections because authenticate() requires a non-empty password.

When the protocol is PeSIT-TLS and the client sends an empty password,
the server now accepts the connection based on the TLS certificate
validation (already performed in VerifyPeerCertificate during the TLS
handshake). The client is identified by its login (PI 3).

Note: A9 (disable pre-connection by default in TLS mode) is deferred
because the PeSIT library's server-side auto-detection has a parsing
bug when pre-connection is absent. This needs a library fix first.

Ref #559

+++ pkg/protocols/modules/pesit/client_transfer.go
@@ -49,7 +49,7 @@ func (c *clientTransfer) configureClient(config *PartnerConfig) *pipeline.Error 
 		c.client.AllowCheckpoints(
 			utils.If(config.CheckpointSize != 0,
 				config.CheckpointSize,
-				c.clientConf.CheckpointSize),
+				uint(c.clientConf.CheckpointSize)),
 			utils.If(config.CheckpointWindow != 0,
 				config.CheckpointWindow,
 				c.clientConf.CheckpointWindow))
